Declaration of functions cRosOutStreamGet() and cRosOutStreamSet() has been moved to the file cros_node.h in order to differentiate these local-printing functions from the ROS log printing functions (declared in cros_log.h)
